Test if menu item is new rather than if node is new.
This is the take care of the case when a user creates new content as a
draft but does not create the menu item at the same time.  If this
occurs, when editting the draft again and a menu item is created it
shows a weight of '0' rather than the set default weight.